Fixes an issue that transaction commit listeners get called multiple times if a transaction is retried.
A thrown exception affects the state of a transaction differently depending on whether it is raised before the transaction is committed or while the transaction is being committed.  If the exception is raised while the transaction is being committed, the low-level transaction object will report it is no longer active.  As such, we can't use the .isActive() method on a transaction to determine if the transaction was committed.  Thus, to ensure that the transaction listeners only get called if the transaction was successfully applied, we have to set the boolean flag committedSuccessfully inside the try block of transactOnce(), where it will only be run if no exceptions are raised from running the transactional Work or from committing the transaction.