Support loading entities with member field of type `Number`.
If you have an entity class with a member field of type `java.lang.Number`, objectify will fail 
to load such an entity from Datastore, throwing `IllegalArgumentException` inside
`NumberTranslatorFactory`. That's because the translator assumes that the type will always
be a subclass of `Number`. This diff fixes the problem by returning the value as-is if the 
requested type is `Number`.