Clarify which versions use what
The Google App Engine API for Java used by Objectify 5 is part of the confusingly named appengine-api-1.0-sdk, but is part of the App Engine SDK for Java only in so far as it is bundled with it. You don't even need the App Engine SDK for Java if you pull that jar in from Maven central.

The Cloud SDK is not used at all by any version of Objectify.